Move Cid and Prefix to their files of the same name
